+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features:
 - Added the classic rare sc sound from old RFU
+- Added the jawbus death sound from old RFU
 
 ### Changes:
 - Merged Flare Timer and Umberella Timer into a single unified Deployables HUD with a per-type toggle;